Convert electoral presenter unit tests
- Move the fixture file used by this test
